Quiet down some compiler warnings.

+++ extras/QPickerTableViewCell.m
@@ -143,9 +143,9 @@ NSString * const QPickerTableViewCellIdentifier = @"QPickerTableViewCell";
 {
     NSArray *componentsValues = [self.pickerElement.valueParser componentsValuesFromObject:value];
     
-    for (int componentIndex = 0; componentIndex < componentsValues.count && componentIndex < _pickerView.numberOfComponents; componentIndex++)
+    for (NSUInteger componentIndex = 0; componentIndex < componentsValues.count && componentIndex < (NSUInteger)_pickerView.numberOfComponents; componentIndex++)
     {
-        id componentValue = [componentsValues objectAtIndex:(NSUInteger) componentIndex];
+        id componentValue = [componentsValues objectAtIndex:componentIndex];
         NSInteger rowIndex = [[self.pickerElement.items objectAtIndex:componentIndex] indexOfObject:componentValue];
         if (rowIndex != NSNotFound) {
             [_pickerView selectRow:rowIndex inComponent:componentIndex animated:YES];
